Embed all code into index file
This is to quickly fix the pipeline. I want to split out into separate
packages in the future, but the LocalStack pipeline is failing.
